Extract celo functions to celo_receipt.go
